// src/api/model/modelPageResultModelComment.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * 我的博客 API
 * 基于 Go + GORM 的博客系统后端
 * OpenAPI spec version: 0.1
 */
import type { ModelComment } from './modelComment';

export interface ModelPageResultModelComment {
  /** 当前页码 */
  current?: number;
  /** 是否有更多 */
  hasMore?: boolean;
  /** 数据列表 */
  list?: ModelComment[];
  /** 每页条数 */
  size?: number;
  /** 总条数 */
  total?: number;
}